pyproject.toml
[tool.poetry]
name = "repl-nix-interactivewebcreator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.46.0"
flask = "^3.0.3"
opencc-python-reimplemented = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
